#261 add functions
