+++ components/services/add/addUserData.ts
@@ -1,0 +1,14 @@
+import { ApiWithMessage } from '../../utils/api';
+import { UserSettings } from '../data/getUserData';
+
+async function addUserData({ body }: { body: UserSettings }) {
+  return await ApiWithMessage(
+    '/api/add/add_user_data',
+    'Saving account settings...',
+    'Account settings saved!',
+    'POST',
+    body
+  );
+}
+
+export { addUserData };
